# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Indonesian KRL Commuter Line Schedule API",
    "title": "Comuline API",
    "contact": {},
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/status": {
      "get": {
        "description": "Returns {\"status\":\"ok\"} when the server is running.",
        "produces": ["application/json"],
        "summary": "Health check",
        "tags": ["system"],
        "responses": {
          "200": {"description": "OK"}
        }
      }
    },
    "/v1/station": {
      "get": {
        "description": "Returns all KRL stations, served from Redis cache when available.",
        "produces": ["application/json"],
        "summary": "List all stations",
        "tags": ["station"],
        "responses": {
          "200": {"description": "OK"}
        }
      }
    },
    "/v1/station/{id}": {
      "get": {
        "description": "Returns a single station by its uppercase ID.",
        "produces": ["application/json"],
        "summary": "Get station by ID",
        "tags": ["station"],
        "parameters": [{"in": "path", "name": "id", "required": true, "type": "string"}],
        "responses": {
          "200": {"description": "OK"},
          "404": {"description": "Not Found"}
        }
      }
    },
    "/v1/schedule/{station_id}": {
      "get": {
        "description": "Returns all train schedules for a station, ordered by departure time.",
        "produces": ["application/json"],
        "summary": "Get schedules by station ID",
        "tags": ["schedule"],
        "parameters": [{"in": "path", "name": "station_id", "required": true, "type": "string"}],
        "responses": {
          "200": {"description": "OK"}
        }
      }
    },
    "/v1/route/{train_id}": {
      "get": {
        "description": "Returns the ordered stop list and route details for a given train ID.",
        "produces": ["application/json"],
        "summary": "Get train route with stop sequence",
        "tags": ["route"],
        "parameters": [{"in": "path", "name": "train_id", "required": true, "type": "string"}],
        "responses": {
          "200": {"description": "OK"},
          "404": {"description": "Not Found"}
        }
      }
    }
  }
}